made all edges of the graph into u64 (make all edges of weight 1)

+++ src/graph.rs
@@ -5,7 +5,7 @@ use ::sprs::CsMat;
 /// Struct that represents a graph
 pub struct Graph{
     /// The CsMat (from sprs) is used to store the graph as a sparse matrix in CSR format
-    pub graph_csr: CsMat<f64>
+    pub graph_csr: CsMat<u64>
 }
 
 impl Graph {
@@ -18,7 +18,7 @@ impl Graph {
     }
 
     /// Set a CsMat for a graph
-    pub fn set_matrix(&mut self, graph_csr: CsMat<f64>) {
+    pub fn set_matrix(&mut self, graph_csr: CsMat<u64>) {
         self.graph_csr = graph_csr;
     }
 
@@ -29,18 +29,18 @@ impl Graph {
     }
 
     /// An iterator over the neighbors of the given vertex.
-    pub fn neighbors(&self, vertex: usize) -> Zip<Cloned<Iter<'_, usize>>, Cloned<Iter<'_, f64>>> {
+    pub fn neighbors(&self, vertex: usize) -> Zip<Cloned<Iter<'_, usize>>, Cloned<Iter<'_, u64>>> {
         let (indices, data) = self.graph_csr.outer_view(vertex).unwrap().into_raw_storage();
         indices.iter().cloned().zip(data.iter().cloned())
     }
 
     /// Insert an edge with two vertices on either ends.
-    pub fn insert(&mut self, vertex1: usize, vertex2: usize, edge_weight: f64) {
+    pub fn insert(&mut self, vertex1: usize, vertex2: usize, edge_weight: u64) {
         self.graph_csr.insert(vertex1, vertex2, edge_weight);
     }
 
     /// Get edge weight for a pair of vertices.
-    pub fn get_edge_weight(&self, vertex1: usize, vertex2: usize) -> Option<f64> {
+    pub fn get_edge_weight(&self, vertex1: usize, vertex2: usize) -> Option<u64> {
         self.graph_csr.get(vertex1, vertex2).cloned()
     }
